requirements.txt
spacy
newspaper3k